requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
psycopg2-binary
pgvector
langchain
langchain-core
langchain-community
langchain_google_genai
langchain-google-community
langchain-postgres
langgraph
huggingface_hub
langchain-huggingface
python-dotenv
redis
pydantic-settings 
psycopg2-binary